+++ C/PythonModule/Neuron.h
@@ -9,7 +9,7 @@ Compute if neuron fires
 @param synapse_activities increase if pre and postsynaptic neurons fire
 @return 1 if fires, otherwise 0
 */
-void compute(struct Synapse **neurons, int num_neurons, int num_synapses_per_neuron, int *neuron_outputs);
+void compute(struct Synapse **neurons, int num_neurons, int num_synapses_per_neuron, float *neuron_outputs);
 
 void read(struct Synapse **neurons, int num_neurons, int num_synapses_per_neuron);
